+++ src/bin/object-store-test/Cargo.toml
@@ -4,6 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-object-store = { version = "0.1.0", path = "../../srv/pager-srv/obliviate/object-store" }
-obliviate-core = { version = "0.1.0", path = "../../srv/pager-srv/obliviate/obliviate-core" }
+object-store = { version = "0.1.0", path = "../../srv/pager-srv/object-store" }
+obliviate-core = { version = "0.1.0", git = "https://github.com/twizzler-operating-system/obliviate.git" }
 rand = "0.8.5"
