+++ src/srv/pager-srv/object-store-old/Cargo.toml
@@ -1,19 +1,15 @@
 [package]
-name = "object-store"
+name = "object-store-old"
 version = "0.1.0"
 edition = "2021"
 
-
 [dependencies]
-chacha20 = { version = "0.9.1", features = ["std", "zeroize"] }
-fatfs = { git = "https://github.com/zphrs/rust-fatfs.git", branch = "zphrs-rwproxy", version = "0.4.0", features = [
+fatfs = { git = "https://github.com/rafalh/rust-fatfs.git", version = "0.4.0", features = [
     "std",
     "log_level_info",
     "lfn",
     "alloc",
 ], default-features = false }
-obliviate-core = { version = "0.1.0", git = "https://github.com/twizzler-operating-system/obliviate.git" }
-rand = "0.8.5"
 twizzler-async = { path = "../../../lib/twizzler-async" }
 twizzler-abi = { path = "../../../lib/twizzler-abi" }
 twizzler-driver = { path = "../../../lib/twizzler-driver" }
