+++ src/srv/pager-srv/object-store-old/src/lib.rs
@@ -1,0 +1,5 @@
+mod disk;
+mod nvme;
+mod object_store;
+
+pub use object_store::{create_object, read_exact, unlink_object, write_all};
