+++ src/srv/pager-srv/object-store-old/src/nvme/controller.rs
@@ -27,6 +27,7 @@ use volatile::map_field;
 use super::{dma::NvmeDmaRegion, requester::NvmeRequester};
 use crate::nvme::dma::NvmeDmaSliceRegion;
 
+#[allow(dead_code)]
 pub struct NvmeController {
     requester: RwLock<Vec<Requester<NvmeRequester>>>,
     admin_requester: RwLock<Option<Arc<Requester<NvmeRequester>>>>,
@@ -439,6 +440,7 @@ impl NvmeController {
         ident.dma_region().with(|ident| ident.clone())
     }
 
+    #[allow(dead_code)]
     pub async fn flash_len(&self) -> usize {
         if let Some(sz) = self.capacity.get() {
             *sz
