--- src/srv/pager-srv/obliviate/object-store/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "object-store"
-version = "0.1.0"
-edition = "2021"
-
-
-[dependencies]
-chacha20 = { version = "0.9.1", features = ["std", "zeroize"] }
-fatfs = { git = "https://github.com/zphrs/rust-fatfs.git", branch = "zphrs-rwproxy", version = "0.4.0", features = [
-    "std",
-    "log_level_info",
-    "lfn",
-    "alloc",
-], default-features = false }
-obliviate-core = { version = "0.1.0", path = "../obliviate-core" }
-rand = "0.8.5"
-twizzler-async = { path = "../../../../lib/twizzler-async" }
-twizzler-abi = { path = "../../../../lib/twizzler-abi" }
-twizzler-driver = { path = "../../../../lib/twizzler-driver" }
-twizzler-object = { path = "../../../../lib/twizzler-object" }
-twizzler-queue = { path = "../../../../lib/twizzler-queue" }
-nvme = { path = "../../../../lib/nvme-rs" }
-async-trait = "0.1.66"
-volatile = "0.5"
-pci-ids = "0.2.4"
